improve dropdowns (#2)
