+++ core-ui/src/components/Predefined/Create/DNSEntries/DNSNameRef.js
@@ -3,9 +3,10 @@ import { useGetList } from 'react-shared';
 import { useTranslation } from 'react-i18next';
 
 import { ResourceForm } from 'shared/ResourceForm/ResourceForm';
+import * as Inputs from 'shared/ResourceForm/components/Inputs';
 
 let memoizedProviders = null;
-export function DNSNameRef({ required, domain, onChange }) {
+export function DNSNameRef({ required, simple, advanced, ...props }) {
   const { t } = useTranslation();
   const url = `/apis/dns.gardener.cloud/v1alpha1/dnsproviders`;
   const { data } = useGetList()(url);
@@ -26,22 +27,20 @@ export function DNSNameRef({ required, domain, onChange }) {
   }));
 
   return (
-    <ResourceForm.FormField
-      required={required}
-      label={t('dnsentries.labels.dns-name')}
-      propertyPath="$.spec.dnsName"
-      setValue={onChange}
-      value={domain}
-      defaultKey={domain}
-      placeholder={t('dnsentries.placeholders.dns-name')}
-      input={props => (
-        <ResourceForm.ComboboxInput
-          key="dns-name-select"
-          options={includedDomains}
-          selectionType="manual"
-          {...props}
-        />
-      )}
-    />
+    <ResourceForm.Wrapper {...props}>
+      <ResourceForm.FormField
+        id="dns-name-ref"
+        required={required}
+        simple={simple}
+        advanced={advanced}
+        label={t('dnsentries.labels.dns-name')}
+        propertyPath="$.spec.dnsName"
+        input={Inputs.ComboboxInput}
+        options={includedDomains}
+        selectionType="manual"
+        key="dns-name-select"
+        placeholder={t('dnsentries.placeholders.dns-name')}
+      />
+    </ResourceForm.Wrapper>
   );
 }
