+++ core-ui/src/components/Predefined/Create/DNSEntries/TargetsRef.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Button, Switch, FormInput } from 'fundamental-react';
+import { Button, Switch, FormInput, ComboboxInput } from 'fundamental-react';
 import classnames from 'classnames';
 import { useGetList, Spinner } from 'react-shared';
 
@@ -101,7 +101,7 @@ const getExternalIPs = loadBalancer => {
 };
 
 let memoizedServices = null;
-export function TargetsRef({ dnsEntry, setTargets, setDnsEntry }) {
+export function TargetsRef({ resource: dnsEntry, setResource: setDnsEntry }) {
   const { t } = useTranslation();
 
   const { data, loading } = useGetList()(`/api/v1/services`);
@@ -113,7 +113,6 @@ export function TargetsRef({ dnsEntry, setTargets, setDnsEntry }) {
   const services = memoizedServices || data;
   if (loading && !services) return <Spinner />;
 
-  const targets = dnsEntry?.spec.targets || [];
   const loadBalancers = services?.filter(
     service =>
       service.spec.type === 'LoadBalancer' &&
@@ -134,7 +133,6 @@ export function TargetsRef({ dnsEntry, setTargets, setDnsEntry }) {
       defaultOpen
       resource={dnsEntry}
       setResource={setDnsEntry}
-      propertyPath="$.spec.targets"
       className="targets-ref"
     >
       <TargetsInput
@@ -147,8 +145,7 @@ export function TargetsRef({ dnsEntry, setTargets, setDnsEntry }) {
             .map(target => target.target)
             .filter(t => !!t)
         }
-        value={targets}
-        setValue={setTargets}
+        propertyPath="$.spec.targets"
         inputs={[
           ({ value, index, setInternalValue }) => (
             <div className="fd-col fd-col-md--3" key={index}>
@@ -193,19 +190,24 @@ export function TargetsRef({ dnsEntry, setTargets, setDnsEntry }) {
             } else {
               return (
                 <div className="fd-col fd-col-md--9" key={index}>
-                  <ResourceForm.ComboboxInput
-                    key={`targets-select-${index}`}
+                  <ComboboxInput
+                    compact
+                    id={'targets-ref'}
+                    ariaLabel="Combobox input"
+                    arrowLabel="Combobox input arrow"
+                    showAllEntries
+                    searchFullString
                     options={IPs}
-                    defaultKey={value?.target}
+                    selectedKey={value?.target}
+                    selectionType="manual"
                     placeholder={t('dnsentries.placeholders.target-a')}
-                    setValue={selected => {
+                    onSelectionChange={(_, selected) => {
                       if (selected.key !== -1) {
                         setValue({ ...value, target: selected.key });
                       } else {
                         setValue({ ...value, target: selected.text });
                       }
                     }}
-                    selectionType="manual"
                   />
                 </div>
               );
