+++ core-ui/src/shared/ResourceForm/ResourceForm.js
@@ -222,5 +222,4 @@ ResourceForm.TextArrayInput = FormComponents.TextArrayInput;
 ResourceForm.K8sNameField = FormComponents.K8sNameField;
 ResourceForm.KeyValueField = FormComponents.KeyValueField;
 ResourceForm.ItemArray = FormComponents.ItemArray;
-ResourceForm.ComboboxInput = FormComponents.ComboboxInput;
 ResourceForm.ComboboxArrayInput = FormComponents.ComboboxArrayInput;
